Implement population determination for faction sectors when generating a new galaxy and generate economic activities on inhabited planets

+++ src/Modules/Galaxy/Domain/Entity/Planet.php
@@ -46,10 +46,13 @@ class Planet extends Place implements SystemUpdatable, \JsonSerializable
 		public Player|null $player = null,
 		#[ORM\Column(type: 'string', length: 45, nullable: true)]
 		public string|null $name = null,
-		#[ORM\Column(type: 'float', options: ['unsigned' => true])]
-		public float       $population = 0.0,
+		/**
+		 * Population unit is million of people
+		 */
+		#[ORM\Column(type: 'integer', options: ['unsigned' => true])]
+		public int $population = 0,
 		#[ORM\Column(type: 'smallint', options: ['unsigned' => true])]
-		public int         $coefResources = 0,
+		public int $coefResources = 0,
 		#[ORM\Column(type: 'smallint', options: ['unsigned' => true])]
 		public int $coefHistory = 0,
 		#[ORM\Column(type: 'smallint', options: ['unsigned' => true, 'default' => 0])]
